pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tokenlens"
version = "0.1.0"
description = "A library for counting tokens across different LLM providers"
authors = [
    {name = "Giri Ramanathan", email = "giri@example.com"}
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# Core dependencies required for basic functionality
dependencies = [
    "tiktoken>=0.5.2",
    "pydantic>=1.10.13",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
]

# Optional dependencies for specific providers
[project.optional-dependencies]
openai = [
    "openai>=1.3.7",
]
anthropic = [
    "anthropic>=0.3.0",
]
google = [
    "google-generativeai>=0.3.0",
]
huggingface = [
    "transformers>=4.30.0",
    "huggingface-hub>=0.16.0",
]
mistral = [
    "mistralai>=0.0.7",
]
stability = [
    "stability-sdk>=0.8.0",
]
media = [
    "torch>=2.0.0",
    "pillow>=10.0.0",
    "moviepy>=1.0.3",
]
api = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "httpx>=0.23.0",
]
# Development dependencies
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]
# Combines all providers
all = [
    "tokenlens[openai]",
    "tokenlens[anthropic]",
    "tokenlens[google]",
    "tokenlens[huggingface]",
    "tokenlens[mistral]",
    "tokenlens[stability]",
    "tokenlens[media]",
    "tokenlens[api]",
]

[tool.setuptools]
packages = ["tokenlens"]
